Extract methods to child classes
