+++ lsp/src/main/java/com/codurance/lsp/ElectricCar.java
@@ -1,21 +1,14 @@
 package com.codurance.lsp;
 
-public class ElectricCar extends Vehicle{
-
-    private static final int BATTERY_FULL = 100;
-    private int batteryLevel;
-
-    @Override
-    public void fillUpWithFuel() {
-        throw new UnsupportedOperationException("It's an electric car");
-    }
-
-    @Override
-    public void chargeBattery() {
-        batteryLevel = BATTERY_FULL;
-    }
+public class ElectricCar extends Vehicle {
+    protected static final int BATTERY_FULL = 100;
+    protected int batteryLevel;
 
     public int batteryLevel() {
         return batteryLevel;
+    }
+
+    public void chargeBattery() {
+        this.batteryLevel = BATTERY_FULL;
     }
 }
