+++ lsp/src/main/java/com/codurance/lsp/FillingStation.java
@@ -2,15 +2,11 @@ package com.codurance.lsp;
 
 public class FillingStation {
 
-    public void refuel(Vehicle vehicle) {
-        if (vehicle instanceof PetrolCar) {
-            vehicle.fillUpWithFuel();
-        }
+    public void refuel(PetrolCar petrolCar) {
+        petrolCar.fillUpWithFuel();
     }
 
-    public void charge(Vehicle vehicle) {
-        if (vehicle instanceof ElectricCar) {
-            vehicle.chargeBattery();
-        }
+    public void charge(ElectricCar electricCar) {
+        electricCar.chargeBattery();
     }
 }
