+++ lsp/src/main/java/com/codurance/lsp/PetrolCar.java
@@ -1,20 +1,14 @@
 package com.codurance.lsp;
 
 public class PetrolCar extends Vehicle {
-  private static final int FUEL_TANK_FULL = 100;
-  private int fuelTankLevel = 0;
+    protected static final int FUEL_TANK_FULL = 100;
+    protected int fuelTankLevel = 0;
 
-  @Override
-  public void fillUpWithFuel() {
-    this.fuelTankLevel = FUEL_TANK_FULL;
-  }
+    public int fuelTankLevel() {
+        return fuelTankLevel;
+    }
 
-  @Override
-  public void chargeBattery() {
-    throw new UnsupportedOperationException("A petrol car cannot be recharged");
-  }
-
-  public int fuelTankLevel() {
-    return fuelTankLevel;
-  }
+    public void fillUpWithFuel() {
+        this.fuelTankLevel = FUEL_TANK_FULL;
+    }
 }
