+++ lsp/src/main/java/com/codurance/lsp/Vehicle.java
@@ -2,7 +2,6 @@ package com.codurance.lsp;
 
 public abstract class Vehicle {
 
-
     private boolean engineStarted = false;
 
     public void startEngine() {
@@ -16,10 +15,4 @@ public abstract class Vehicle {
     public void stopEngine() {
         this.engineStarted = false;
     }
-
-    public abstract void fillUpWithFuel();
-
-    public abstract void chargeBattery();
-
-
 }
