+++ lsp/src/test/java/com/codurance/lsp/VehicleTest.java
@@ -1,0 +1,34 @@
+package com.codurance.lsp;
+
+import org.junit.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class VehicleTest {
+    @Test
+    public void start_engine() {
+        Vehicle vehicle = new TestableVehicle();
+
+        vehicle.startEngine();
+
+        assertThat(vehicle.engineIsStarted())
+                .isTrue();
+
+    }
+
+    @Test
+    public void stop_engine() {
+        Vehicle vehicle = new TestableVehicle();
+
+        vehicle.startEngine();
+        vehicle.stopEngine();
+
+        assertThat(vehicle.engineIsStarted())
+                .isFalse();
+    }
+
+
+    public static class TestableVehicle extends Vehicle {
+    }
+
+}
